Add top-level CMakeLists.txt for the entire BBB repository
Modified the BBB_cci_mpf/sw CMake rules to support either independent
compilation with the root in BBB_cci_mpf/sw or compilation as part of the whole
BBB repository.  MPF documentation is now compiled with the "mpf_doc" target.
The top-level build defines the "doc" target, which can depend on all
sub-projects in the entire repository.