MPF/VTP now can pin and translate pages automatically (#37)
VTP can now be configured in two modes. The old version has a hardware page table
walker. The new, default, version traps to software when a page translation is not found
in the hardware-side caches. The software is actually faster under load since the
original hardware version is not pipelined. The software is also capable of pinning
pages on demand, making it possible to pass arbitrary virtual addresses for use in the
FPGA without first calling fpgaPrepareBuffer().