Add the mmu_monitor driver.
The mmu_monitor driver creates a device file: /dev/mmu_monitor. When opened, the
driver registers a kernel MMU notifier. The notifier tracks changes to the MMU
mapping for the process. The mmu_monitor driver forwards MMU updates as events
to user space. These can be used to track changes to memory regions that are
pinned for access by an FPGA. MPF's virtual to physical (VTP) module uses this
driver when it is available.

The mmu_monitor driver is a temporary solution, enabling a fully automatic
MPF/VTP service on current hardware and kernels. Future systems will use built-in
services, such as PCIe ATS (address translation service), to manage FPGA shared
virtual memory.