Deduplicate back-to-back MPF/VTP L1 misses.
L2 VTP bandwidth is shared by all pipelines, so forwarding duplicate requests
may affect performance. This is especially true on L2 misses, since the page
table walkers also don't deduplicate requests. The new code detects back-to-back
L2 requests to the same page, forwards only the first to the L2 lookup, and
uses the response to represent all of the requests.