Use GNUInstallDirs to specify MPF install targets
The GNUInstallDirs provides standard installation directories for GNU
software (https://cmake.org/cmake/help/v3.0/module/GNUInstallDirs.html).
This modifies the install command to use these directories, which
allows the targets to be overwritten if needed (i.e. for OpenCL MMD).

Also update the minimum CMake version to be consistent with OPAE
requirements.  This ensures the GNUInstallDirs feature is available.